% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex-extract-install.R
\name{read_tex}
\alias{read_tex}
\title{read tex file into R}
\usage{
read_tex(path, ...)
}
\arguments{
\item{path}{file path}

\item{...}{extra options to pass to readChar}
}
\value{
text string
}
\description{
read tex file into R
}
\examples{
\dontrun{
tex_chars <- read_tex("path/to/file.tex")
}
}
