% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tex-example.R
\docType{data}
\name{tex_example}
\alias{tex_example}
\title{Example TeX template from PLoS ONE}
\format{An object of class \code{character} of length 1.}
\usage{
tex_example
}
\description{
This is an example tex string from the PLoS ONE template:
  https://journals.plos.org/plosone/s/latex
}
\examples{
tex_preamble <- tex_extract_preamble(tex_example)
}
\keyword{datasets}
