% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex-extract-install.R
\name{tex_extract_pkgs}
\alias{tex_extract_pkgs}
\title{Extract packages from tex string}
\usage{
tex_extract_pkgs(tex_string)
}
\arguments{
\item{tex_string}{text string}
}
\value{
text string of package names
}
\description{
Extract packages from tex string
}
\examples{
tex_preamble <- tex_extract_preamble(tex_example)
tex_extract_pkgs(tex_preamble)
tex_extract_pkgs(tex_example)
}
