% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex-extract-install.R
\name{tex_extract_preamble}
\alias{tex_extract_preamble}
\title{Extract preamble from tex string}
\usage{
tex_extract_preamble(tex_string)
}
\arguments{
\item{tex_string}{tex string, could be read in with read_tex}
}
\value{
text string containing all text before `\\begin{document}`
}
\description{
Extract preamble from tex string
}
\examples{
\dontrun{
tex_preamble <- tex_extract_preamble(tex_example)
}
}
