% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tex-extract-install.R
\name{tex_install_pkgs}
\alias{tex_install_pkgs}
\title{Install tex packages}
\usage{
tex_install_pkgs(tex_pkg_string, ...)
}
\arguments{
\item{tex_pkg_string}{string}

\item{...}{extra options to pass to `tinytex::tlmgr_install()` .}
}
\value{
nothing, this installs packages
}
\description{
Install tex packages using `tinytex::tlmgr_install()`
}
\examples{
\dontrun{
# in a pipeline
read_tex(tex_path) \%>\%
  tex_extract_preamble() \%>\%
  tex_extract_pkgs() \%>\%
  tex_install_pkgs()
  }
}
